Refactor: Move semantic analysis of call-based Enum to a new module
